+++ CHANGELOG.md
@@ -34,16 +34,18 @@ PositionGroup.alter()
     - Checks `IntervalList` and externals tables. #1002
     - Ends early if called on empty table. #1055
 - Allow mixin tables with parallelization in `make` to run populate with
-    `processes > 1` #1001, #1052
+    `processes > 1` #1001, #1052, #1068
 - Speed up fetch_nwb calls through merge tables #1017
 - Allow `ModuleNotFoundError` or `ImportError` for optional dependencies #1023
 - Ensure integrity of group tables #1026
 - Convert list of LFP artifact removed interval list to array #1046
-- Merge duplicate functions in decoding and spikesorting #1050, #1053, #1058
+- Merge duplicate functions in decoding and spikesorting #1050, #1053, #1062,
+    #1069
 - Revise docs organization.
     - Misc -> Features/ForDevelopers. #1029
     - Installation instructions -> Setup notebook. #1029
 - Migrate SQL export tools to `utils` to support exporting `DandiPath` #1048
+- Add tool for checking threads for metadata locks on a table #1063
 - Disable populate transaction protection for long-populating tables #1066
 
 ### Pipelines